placed joystick support classes into their own source files
