+++ combat_platformer/src/combat_platformer/input/__init__.py
@@ -1,2 +1,5 @@
 from joystick_buttons import *
 from move import *
+from input_manager import *
+from joystick_state import *
+from joystick_manager import *
