+++ combat_platformer/src/combat_platformer/input/move.py
@@ -5,7 +5,7 @@ class Move(object):
         Move(name,button_sequence,is_submove,callback)
         Inputs:
             - name             : String containing the name for the move
-            - button_sequence  : List containing JoystickButton or Keyboard bitmasks for activating the move
+            - button_sequence  : List containing a JoystickButton list that activates the move
             - is_submove : True | False indicating if this move is part of a larger move
             - callback : The function that will be executed when the execute() method is invoked
     """
@@ -14,7 +14,7 @@ class Move(object):
         
         self.name = name
         self.button_sequence = button_sequence;
-        self.callback = callback if (callback != None) else (lambda : "Move %s executed"%(name))
+        self.callback = callback if (callback != None) else (lambda : "Move '%s' executed"%(name))
         self.is_submove = is_submove
         
     def match(self,sequence):
